refactor: add event listener on each thambnails.
