+++ js/comments.js
@@ -1,9 +1,7 @@
-import { getComment } from './node-maker.js';
-
-
 const COMMENTS_SHOW_STEP = 5;
 
 
+const commentTemplate = document.querySelector('#comment').content.querySelector('.social__comment');
 const bigPicture = document.querySelector('.big-picture');
 const socialComments = bigPicture.querySelector('.social__comments');
 const socialCommentTotalCount = bigPicture.querySelector('.social__comment-total-count');
@@ -13,6 +11,25 @@ const commentsLoader = bigPicture.querySelector('.comments-loader');
 
 let commentsLimitCurrent = COMMENTS_SHOW_STEP;
 let comments = [];
+
+
+/**
+ * Создаёт разметку комментария.
+ * @param {object} commentData Данные комментария.
+ * @returns {object} Комментарий.
+*/
+const getComment = (commentData) => {
+  const comment = commentTemplate.cloneNode(true);
+  const commentImg = comment.querySelector('img');
+  const commentMessage = comment.querySelector('p');
+
+  comment.dataset.commentId = commentData.id;
+  commentImg.src = commentData.avatar;
+  commentImg.alt = commentData.name;
+  commentMessage.textContent = commentData.message;
+
+  return comment;
+};
 
 
 /**
