+++ js/gallery.js
@@ -1,19 +1,46 @@
 import { mockPhotos } from './mock-photos.js';
-import { getThumbnail } from './node-maker.js';
 import { openPicture } from './big-picture.js';
 
 
 const picturesContainer = document.querySelector('.pictures');
+const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
 
 /**
- * Вставляет превью фотографий на страницу.
- * @param {object} photos Список фотографий.
+ * Создаёт разметку превью изображения.
+ * @param {object} pictureData Данные изображения.
+ * @returns {object} Превью изображения.
+*/
+const getThumbnail = (pictureData) => {
+  const thumbnail = thumbnailTemplate.cloneNode(true);
+  // thumbnail.dataset.photoId = photoData.id;
+
+  const thumbnailPicture = thumbnail.querySelector('.picture__img');
+  thumbnailPicture.src = pictureData.url;
+  thumbnailPicture.alt = pictureData.description;
+
+  const thumbnailCommentsCount = thumbnail.querySelector('.picture__comments');
+  thumbnailCommentsCount.textContent = pictureData.comments.length;
+
+  const thumbnailLikesCount = thumbnail.querySelector('.picture__likes');
+  thumbnailLikesCount.textContent = pictureData.likes;
+
+  thumbnail.addEventListener('click', () => {
+    openPicture(pictureData);
+  });
+
+  return thumbnail;
+};
+
+
+/**
+ * Вставляет превью изображений на страницу.
+ * @param {object} pictures Список изображений.
  */
-const insertThumbnails = (photos) => {
+const insertThumbnails = (pictures) => {
   const thumbnailsFragment = document.createDocumentFragment();
 
-  photos.forEach((photoData) => {
+  pictures.forEach((photoData) => {
     const thumbnail = getThumbnail(photoData);
     thumbnailsFragment.append(thumbnail);
   });
@@ -21,15 +48,5 @@ const insertThumbnails = (photos) => {
   picturesContainer.append(thumbnailsFragment);
 };
 
-
-// Отслеживает нажатие на миниатюру фотографии.
-picturesContainer.addEventListener('click', (evt) => {
-  const currentPhoto = evt.target.closest('.picture');
-
-  if (currentPhoto) {
-    evt.preventDefault();
-    openPicture(mockPhotos.find((photo) => photo.id === Number(currentPhoto.dataset.photoId)));
-  }
-});
 
 insertThumbnails(mockPhotos);
